requirement.txt
selenium
pyyaml
lxml